Moved creation of MWAAEnviroment from base_stack to standard_pipeline stack
