+++ sdlf-ddk-lightweight/python/data_lake/pipelines/sdlf_base_stack.py
@@ -21,7 +21,7 @@ import aws_cdk as cdk
 import aws_cdk.aws_iam as iam
 import aws_cdk.aws_lambda as lmbda
 import aws_cdk.aws_ssm as ssm
-from aws_ddk_core import BaseStack, MWAAEnvironment
+from aws_ddk_core import BaseStack
 from constructs import Construct
 
 from ..foundations import FoundationsStack
@@ -80,7 +80,7 @@ class SDLFBaseStack(BaseStack):
 
         dataset_names: set[str] = set()
         pipelines: Dict[str, SDLFPipeline] = {}
-        self._mwaa_env = None
+        
         # loop through values in parameters.json and create the necessary resources for each pipeline
         for customer_config in customer_configs:
             dataset = customer_config["dataset"]
@@ -92,10 +92,6 @@ class SDLFBaseStack(BaseStack):
                 "orchestration", "sfn"
             ).lower()
 
-            
-            if(orchestration == "mwaa") and (not self._mwaa_env or f"{self._resource_prefix}-{team}-{self._environment_id}" not in self._mwaa_env.name):
-            # creates "MWAA environment" if orchestration is enabled through MWAA
-                self._mwaa_env = self._create_mwaa_env(team, pipeline_type)
             # PIPELINE CREATION
             pipeline: SDLFPipeline
             pipeline_name = f"{team}-{pipeline_type}"
@@ -107,7 +103,6 @@ class SDLFBaseStack(BaseStack):
                         environment_id=self._environment_id,
                         resource_prefix=self._resource_prefix,
                         team=team,
-                        mwaa_env =self._mwaa_env,
                         orchestration=orchestration,
                         foundations_stage=self._foundations_stage,
                         wrangler_layer=self._wrangler_layer,
@@ -143,53 +138,7 @@ class SDLFBaseStack(BaseStack):
                 pipeline.register_dataset(
                     dataset, config=customer_config.get("config", {})
                 )
-    def _create_mwaa_env(self,team,pipeline) -> None:
-        # mwaaa get all configs dynamically through ddk.json
-        data_lake_mwaa_env = MWAAEnvironment(
-            self,
-            id=f"{self._resource_prefix}-{team}-{self._environment_id}-mwaa-environment",
-            name=f"{self._resource_prefix}-{team}-{self._environment_id}-mwaa-environment",
-            airflow_configuration_options={'core.dag_run_conf_overrides_params':True} ,
-            vpc_cidr="10.44.0.0/16 ",
-            environment_class="mw1.small",
-            max_workers=1,
-            dag_s3_path="dags",
-            dag_files=[
-                os.path.join(f"{Path(__file__).parents[1]}", f"src/dags/{team}")
-            ],  
-
-            additional_policy_statements = [
-                iam.PolicyStatement(
-                    effect=iam.Effect.ALLOW,
-                    actions=["lambda:InvokeFunction"],
-                    resources=[
-                        f"arn:aws:lambda:{cdk.Aws.REGION}:{cdk.Aws.ACCOUNT_ID}"
-                        + f":function:{self._resource_prefix}-*"
-                    ],
-                ),
-                iam.PolicyStatement(
-                    effect=iam.Effect.ALLOW,
-                    actions=[
-                        "logs:CreateLogStream",
-                        "logs:CreateLogGroup",
-                        "logs:PutLogEvents",
-                        "logs:GetLogEvents",
-                        "logs:GetLogRecord",
-                        "logs:GetLogGroupFields",
-                        "logs:GetQueryResults",
-                        "logs:DescribeLogGroups"
-
-                    ],
-                    resources=[
-                        f"arn:aws:logs:{cdk.Aws.REGION}:{cdk.Aws.ACCOUNT_ID}"
-                        + f":log-group:airflow-{self._resource_prefix}-*"
-                    ],
-                ),
-
-            ],
-        )
-
-        return data_lake_mwaa_env.mwaa_environment    
+        
     def _create_wrangler_layer(self):
         wrangler_layer_version = lmbda.LayerVersion.from_layer_version_arn(
             self,
